fix build: Int32 crypto status, non-throwing synchronizeFile, async-safe NSLock
- StreamDecryptor: kCCUnimplemented is Int, cryptoFailure takes Int32
- SpotifyCDNClient: synchronizeFile() is non-throwing; drop misplaced try
- DownloadManager: extract lock-guarded state access into sync helpers
  (beginDownloadIfIdle / recordDownloadedFile) so NSLock is never touched
  from async contexts (Swift 6 mode)

+++ Sources/EeveeSpotify/Download/Core/StreamDecryptor.swift
@@ -203,7 +203,8 @@ final class StreamCTREncryptor {
     /// exactly the decrypted bytes (one unavoidable `Data` copy per chunk).
     func decrypt(_ chunk: Data) throws -> Data {
         guard let cryptor = cryptor else {
-            throw StreamDecryptor.DecryptError.cryptoFailure(kCCUnimplemented)
+            // kCCUnimplemented is an Int constant; cryptoFailure takes Int32.
+            throw StreamDecryptor.DecryptError.cryptoFailure(Int32(kCCUnimplemented))
         }
         guard !chunk.isEmpty else { return Data() }
